\name{tutorial.2}
\docType{data}
\alias{tutorial.2}
\title{Real data set for BaSAR tutorial}
\description{
This data set is one time series and its corresponding time vector, for usage together with the BaSAR tutorial.
}
     \usage{tutorial.2}
     \format{A table with one column for time, and one column with experimentally measured calcium oscillations, separated by commas.}
     \keyword{datasets}