\name{tutorial}
\docType{data}
\alias{tutorial}
\title{Simulated data set for BaSAR tutorial}
\description{
	This data set is a small colllection of simulated time series, for usage together with the BaSAR tutorial.
}
     \usage{tutorial}
     \format{A table with one column for time, and four simulated time series, separated by commas.}
     \keyword{datasets}